+++ test/blackbox-tests/test-cases/github7806.t
@@ -1,9 +1,7 @@
-  $ dune init project name $PWD 2>&1 | head -n 8
+  $ dune init project name $PWD
   Entering directory 'name'
-  Internal error, please report upstream including the contents of _build/log.
-  Description:
-    ("[as_in_source_tree_exn] called on something not in source tree",
-    { t =
-        External
-          "$TESTCASE_ROOT"
-    })
+  Error: Path
+  $TESTCASE_ROOT
+  is not relative.
+  Leaving directory 'name'
+  [1]